[SPARK-6785][SQL] move test case for DateTimeUtils to DateTimeUtilsSuite
