+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/MiscFunctionsSuite.scala
@@ -17,11 +17,7 @@
 
 package org.apache.spark.sql.catalyst.expressions
 
-import java.sql.Date
-import java.text.SimpleDateFormat
-
 import org.apache.spark.SparkFunSuite
-import org.apache.spark.sql.catalyst.util.DateTimeUtils
 import org.apache.spark.sql.types.BinaryType
 
 class MiscFunctionsSuite extends SparkFunSuite with ExpressionEvalHelper {
@@ -32,42 +28,4 @@ class MiscFunctionsSuite extends SparkFunSuite with ExpressionEvalHelper {
       "6ac1e56bc78f031059be7be854522c4c")
     checkEvaluation(Md5(Literal.create(null, BinaryType)), null)
   }
-
-  test("SPARK-6785: java date conversion before and after epoch") {
-    def checkFromToJavaDate(d1: Date): Unit = {
-      val d2 = DateTimeUtils.toJavaDate(DateTimeUtils.fromJavaDate(d1))
-      assert(d2.toString === d1.toString)
-    }
-
-    val df1 = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
-    val df2 = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss z")
-
-    checkFromToJavaDate(new Date(100))
-
-    checkFromToJavaDate(Date.valueOf("1970-01-01"))
-
-    checkFromToJavaDate(new Date(df1.parse("1970-01-01 00:00:00").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1970-01-01 00:00:00 UTC").getTime))
-
-    checkFromToJavaDate(new Date(df1.parse("1970-01-01 00:00:01").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1970-01-01 00:00:01 UTC").getTime))
-
-    checkFromToJavaDate(new Date(df1.parse("1969-12-31 23:59:59").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1969-12-31 23:59:59 UTC").getTime))
-
-    checkFromToJavaDate(Date.valueOf("1969-01-01"))
-
-    checkFromToJavaDate(new Date(df1.parse("1969-01-01 00:00:00").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1969-01-01 00:00:00 UTC").getTime))
-
-    checkFromToJavaDate(new Date(df1.parse("1969-01-01 00:00:01").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1969-01-01 00:00:01 UTC").getTime))
-
-    checkFromToJavaDate(new Date(df1.parse("1989-11-09 11:59:59").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1989-11-09 19:59:59 UTC").getTime))
-
-    checkFromToJavaDate(new Date(df1.parse("1776-07-04 10:30:00").getTime))
-    checkFromToJavaDate(new Date(df2.parse("1776-07-04 18:30:00 UTC").getTime))
-  }
-
 }
